dashboard/requirements.txt
docker-py
Flask